requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.4
openai==0.27.8
google-generativeai==0.3.1
anthropic==0.5.0
librosa==0.10.0
pydub==0.25.1
eventlet==0.33.3
gunicorn==20.1.0
python-engineio==4.3.4
python-socketio==5.7.2 